+++ app/Http/Controllers/TweetController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateTweetRequest;
 use App\Models\Tweet;
 use App\Services\TweetService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TweetController extends Controller
 {
@@ -22,6 +23,7 @@ class TweetController extends Controller
    */
   public function index()
   {
+    Gate::authorize('viewAny', Tweet::class);
     $tweets = $this->tweetService->allTweets();
     return view('tweets.index', compact('tweets'));
   }
@@ -31,6 +33,7 @@ class TweetController extends Controller
    */
   public function create()
   {
+    Gate::authorize('create', Tweet::class);
     return view('tweets.create');
   }
 
@@ -39,8 +42,8 @@ class TweetController extends Controller
    */
   public function store(StoreTweetRequest $request)
   {
+    Gate::authorize('create', Tweet::class);
     $this->tweetService->createTweet($request);
-
     return redirect()->route('tweets.index');
   }
 
@@ -49,6 +52,7 @@ class TweetController extends Controller
    */
   public function show(Tweet $tweet)
   {
+    Gate::authorize('view', $tweet);
     $tweet->load('comments');
     return view('tweets.show', compact('tweet'));
   }
@@ -58,6 +62,7 @@ class TweetController extends Controller
    */
   public function edit(Tweet $tweet)
   {
+    Gate::authorize('update', $tweet);
     return view('tweets.edit', compact('tweet'));
   }
 
@@ -66,8 +71,8 @@ class TweetController extends Controller
    */
   public function update(UpdateTweetRequest $request, Tweet $tweet)
   {
+    Gate::authorize('update', $tweet);
     $this->tweetService->updateTweet($request, $tweet);
-
     return redirect()->route('tweets.show', $tweet);
   }
 
@@ -76,8 +81,8 @@ class TweetController extends Controller
    */
   public function destroy(Tweet $tweet)
   {
+    Gate::authorize('delete', $tweet);
     $this->tweetService->deleteTweet($tweet);
-
     return redirect()->route('tweets.index');
   }
 }
